feat(web): Extract bottle resolver
Move the photo scan, identification, result review, start-over, and search fallback UI out of the tasting route into a reusable resolver component.

Keep /addTasting responsible for tasting-specific tag loading and submit behavior so scan-to-tasting and pending image handling stay unchanged.
